ch12/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
Django==4.1.3
pip==22.2.2
setuptools==65.5.0
sqlparse==0.4.3
tzdata==2022.6
wheel==0.37.1
wincertstore==0.2
pkgconfig==1.5.5
mysqlclient==2.2.0
Markdown==3.4.4
django-simple-captcha==0.5.18
pymongo==4.5.0
plotly==5.16.1
numpy==1.25.2
django-registration-redux==2.12
django-allauth==0.55.2
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
idna==3.4
oauthlib==3.2.1
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
types-cryptography==3.3.23
urllib3==1.26.12
django-filer==3.0.6
django-mptt==0.14.0
django-shopping-cart==0.1
django-paypal==2.0